Moved API instantiation for gpV1 to its own file
