+++ src/hooks/useBetterAddTokenModal.tsx
@@ -5,7 +5,7 @@ import { Deferred, createDeferredPromise } from 'utils'
 import { TokenDetails } from 'types'
 import TokenImg from '../components/common/TokenImg'
 import styled from 'styled-components'
-import { tokenListApi } from 'api'
+import { tokenListApi } from 'api/gpV1'
 import { TokenFromExchange } from 'services/factories'
 import { unstable_batchedUpdates as batchUpdates } from 'react-dom'
 import useSafeState from './useSafeState'
