+++ src/hooks/useEnableToken.ts
@@ -1,7 +1,7 @@
 import { assert } from '@gnosis.pm/dex-js'
 
 import { TokenBalanceDetails, TxOptionalParams, Receipt } from 'types'
-import { erc20Api, depositApi } from 'api'
+import { erc20Api, depositApi } from 'api/gpV1'
 
 import useSafeState from './useSafeState'
 import { useWalletConnection } from './useWalletConnection'
