+++ src/hooks/useOrders.ts
@@ -2,7 +2,7 @@ import { useEffect, useCallback, useRef } from 'react'
 import { unstable_batchedUpdates as batchedUpdates } from 'react-dom'
 
 // API + Reducer/Actions
-import { exchangeApi } from 'api'
+import { exchangeApi } from 'api/gpV1'
 import { getTokenFromExchangeById } from 'services'
 import { overwriteOrders, updateOffset, updateOrders } from 'reducers-actions/orders'
 // Hooks
