+++ src/hooks/usePendingOrders.ts
@@ -1,6 +1,6 @@
 import { useEffect, useMemo } from 'react'
 // API
-import { web3 } from 'api'
+import { web3 } from 'api/gpV1'
 // Hooks
 import useGlobalState from './useGlobalState'
 import useSafeState from './useSafeState'
