+++ src/hooks/useTokenList.ts
@@ -3,7 +3,7 @@ import { TokenDetails } from 'types'
 import useSafeState from './useSafeState'
 import { getTokens, subscribeToTokenList } from 'services'
 import { EMPTY_ARRAY } from 'const'
-import { tokenListApi } from 'api'
+import { tokenListApi } from 'api/gpV1'
 
 export interface UseTokenListParams {
   networkId?: number
