+++ src/hooks/useWalletConnection.ts
@@ -1,4 +1,4 @@
-import { walletApi } from 'api'
+import { walletApi } from 'api/gpV1'
 import { useEffect, useMemo } from 'react'
 import { Command, Network } from 'types'
 import useSafeState from './useSafeState'
