+++ src/hooks/useWithdrawTokens.ts
@@ -1,6 +1,6 @@
 import { TokenBalanceDetails, TxOptionalParams, Receipt } from 'types'
 import { assert } from '@gnosis.pm/dex-js'
-import { depositApi } from 'api'
+import { depositApi } from 'api/gpV1'
 import { useWalletConnection } from './useWalletConnection'
 import useSafeState from './useSafeState'
