+++ webpack.config.js
@@ -224,6 +224,7 @@ module.exports = ({ stats = false } = {}) => ({
       // AUTOCONNECT: only applies for mock implementation
       AUTOCONNECT: 'true',
       PRICE_ESTIMATOR_URL: process.env.PRICE_ESTIMATOR_URL || (isProduction && 'production') || 'develop',
+      APP: null,
       APP_ID: null,
       INFURA_ID: null,
       WALLET_CONNECT_BRIDGE: null,
